feat: bolt12 trait
