+++ crates/cdk-lnbits/src/lib.rs
@@ -12,14 +12,12 @@ use async_trait::async_trait;
 use axum::Router;
 use cdk::amount::{to_unit, Amount, MSAT_IN_SAT};
 use cdk::cdk_lightning::{
-    self, Bolt12PaymentQuoteResponse, CreateInvoiceResponse, CreateOfferResponse, MintLightning,
-    PayInvoiceResponse, PaymentQuoteResponse, Settings, WaitInvoiceResponse,
+    self, CreateInvoiceResponse, MintLightning, PayInvoiceResponse, PaymentQuoteResponse, Settings,
+    WaitInvoiceResponse,
 };
 use cdk::mint::types::PaymentRequest;
 use cdk::mint::FeeReserve;
-use cdk::nuts::{
-    CurrencyUnit, MeltQuoteBolt11Request, MeltQuoteBolt12Request, MeltQuoteState, MintQuoteState,
-};
+use cdk::nuts::{CurrencyUnit, MeltQuoteBolt11Request, MeltQuoteState, MintQuoteState};
 use cdk::util::unix_time;
 use cdk::{mint, Bolt11Invoice};
 use error::Error;
@@ -334,35 +332,6 @@ impl MintLightning for LNbits {
         };
 
         Ok(pay_response)
-    }
-
-    async fn get_bolt12_payment_quote(
-        &self,
-        _melt_quote_request: &MeltQuoteBolt12Request,
-    ) -> Result<Bolt12PaymentQuoteResponse, Self::Err> {
-        Err(Error::UnsupportedMethod.into())
-    }
-
-    /// Pay a bolt12 offer
-    async fn pay_bolt12_offer(
-        &self,
-        _melt_quote: mint::MeltQuote,
-        _amount: Option<Amount>,
-        _max_fee_amount: Option<Amount>,
-    ) -> Result<PayInvoiceResponse, Self::Err> {
-        Err(Error::UnsupportedMethod.into())
-    }
-
-    /// Create bolt12 offer
-    async fn create_bolt12_offer(
-        &self,
-        _amount: Option<Amount>,
-        _unit: &CurrencyUnit,
-        _description: String,
-        _unix_expiry: u64,
-        _single_use: bool,
-    ) -> Result<CreateOfferResponse, Self::Err> {
-        Err(Error::UnsupportedMethod.into())
     }
 }
